requirements.txt
# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiogram==3.2.0
aiohttp==3.9.1

# Database
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Cache and async processing
redis==5.0.1

# AI/ML APIs
openai==1.3.7
replicate==0.22.0

# Payment systems
yookassa==3.0.0

# File storage
cloudinary==1.36.0

# Environment and config
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database (SQLite support)
aiosqlite==0.19.0

# Utilities
pillow==10.1.0
httpx==0.25.2
loguru==0.7.2
